created many more classes and slightly re-organized infrastructure; promoted ints and doubles to decimals for consistency.

+++ Training/Proctologist/Models/Statistic.cs
@@ -5,7 +5,7 @@
             set;
         }
 
-        public double Value {
+        public decimal Value {
             get;
             set;
         }
